* Return a thread ID, not an error code, as per Chris' suggestion ;) * pthread_t is an unsigned long, not just a measly int
llvm-svn: 12999

+++ llvm/runtime/libthread/threadlib.h
@@ -1,8 +1,8 @@
 #ifndef THREADLIB_H
 #define THREADLIB_H
 
-int __llvm_thread_start(void*(*fn)(void*), void* arg);
+unsigned long __llvm_thread_start(void*(*fn)(void*), void* arg);
 
-void __llvm_thread_join(int thread_id);
+void __llvm_thread_join(unsigned long thread_id);
 
 #endif
